Improve devnet bootstrap and txnstream server
